footer changed successfully and added new features to homepage
the footer is finally sticky to the bottom of the page and i also added new features to the homepage, it waslooking empty and i needed to fill in the empty space, next the images will be added. i also added div on the pages that needed to be added on for the css features to function properly.