the website has been completed
i have added all the information required to complete the website. i added all the animations to the CSS. File, i added 4images including background cover and also edited the header of the website to suit the overall design and feel. all i need to do is complete the Readme. File